+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - LVM: Removed trailing 'a|.*|' from LVM filter
 
+### Fixed
+
+- Do not update satellites that are evicted
+
 ## [1.26.0] - 2024-01-29
 
 ## [1.26.0-rc.1] - 2024-01-22